+++ src/sap/SapAuthType.ts
@@ -1,3 +1,3 @@
-export type SapAuthType = 'basic' | 'jwt' | 'saml';
+export type SapAuthType = 'basic' | 'jwt' | 'saml' | 'certificate' | 'kerberos';
 
 export type SapConnectionType = 'http' | 'rfc';